Fix AOS_SetupSolution from master

+++ src/Structs/Comp.java
@@ -3,6 +3,6 @@ package Structs;
 public class Comp {
     public Double[] dz;
     public Double[] dzsum;
-    public Double[] layer;
+    public Integer[] layer;
     public Double[] th_fc;
 }
